requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
python-dotenv==1.0.0
openai==1.3.0
firebase-admin==6.2.0
pymongo==4.6.0
google-cloud-firestore==2.13.1